Migrate core and editor Portals to stable endpoint IDs (#461)

+++ src/BooleanWorld/render/include/PortalLight.h
@@ -34,7 +34,7 @@ inline constexpr float PortalLightTransformTolerance = 1e-5f;
 struct PortalLightEndpointKey {
   uint32_t layerId{};
   uint32_t pairId{};
-  uint8_t endpointId{};
+  uint32_t endpointId{};
 
   auto operator<=>(PortalLightEndpointKey const&) const = default;
 };
