+++ src/BooleanWorld/render/include/PortalView.h
@@ -19,7 +19,7 @@
 struct PortalEndpointKey {
   uint32_t layerId{};
   uint32_t pairId{};
-  uint8_t endpointId{};
+  uint32_t endpointId{};
 
   auto operator<=>(PortalEndpointKey const&) const = default;
 };
